requirements.txt
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
ollama>=0.1.0
python-dotenv>=1.0.0

